Backend/requirements.txt
# Python version requirement
# Requires Python >=3.8,<3.12
tensorflow-cpu==2.15.0
Flask==2.3.2
flask-cors==4.0.0
numpy==1.24.3
scikit-learn==1.3.0


